#8078 - Refactoring
