+++ sormas-ui/src/main/java/de/symeda/sormas/ui/vaccination/list/VaccinationList.java
@@ -18,7 +18,6 @@ package de.symeda.sormas.ui.vaccination.list;
 import java.util.List;
 import java.util.function.Function;
 
-import com.vaadin.shared.Registration;
 import com.vaadin.ui.Label;
 
 import de.symeda.sormas.api.Disease;
@@ -27,7 +26,6 @@ import de.symeda.sormas.api.i18n.I18nProperties;
 import de.symeda.sormas.api.vaccination.VaccinationListEntryDto;
 import de.symeda.sormas.ui.utils.PaginationList;
 import de.symeda.sormas.ui.utils.components.sidecomponent.event.EditSideComponentFieldEvent;
-import de.symeda.sormas.ui.utils.components.sidecomponent.event.EditSideComponentFieldEventListener;
 
 public class VaccinationList extends PaginationList<VaccinationListEntryDto> {
 
@@ -66,12 +64,5 @@ public class VaccinationList extends PaginationList<VaccinationListEntryDto> {
 				e -> fireEvent(new EditSideComponentFieldEvent(listEntry)));
 			listLayout.addComponent(listEntry);
 		}
-	}
-
-	public Registration addSideComponentFieldEditEventListener(EditSideComponentFieldEventListener editSideComponentFieldEventListener) {
-		return addListener(
-			EditSideComponentFieldEvent.class,
-			editSideComponentFieldEventListener,
-			EditSideComponentFieldEventListener.ON_EDIT_SIDE_COMPONENT_FIELD_METHOD);
 	}
 }
